updated internal method and printing syntax
all methods internal to a class or module are now prefixed with an
underscore. This will help when docstrings are added for each module
API.

Additionally many of the error message print statements have been
updated to use f"{var}" syntax so that more instructive error messages
can be printed, formatted more nicely.