Allow functions with randomvars to fail sometimes
Testing a function which produces a random variable is problematic
because for some inputs, it will fail tests, even if the function is
correct. Now if the expected output is a randomvar, then we test to make
sure that the function passes 50% of the tests. That way if the student
flukes a right answer, they still get an error message, but if their
code 'flukes' an incorrect answer some of the time (As is expected) they
still pass the tests.

In either case, the p-value for the failing test gets printed, and so
they can judge for themselves whether things look ok.