Fixed failing tests of randomvars with pvalues
When testing functions that produce random variables, the function is
called multiple times, and the p-value is calculated in order to
determine if the random variable satisfies a hypothesis test.

Previously, we stored only the minimum p-value obtained, meaning that
even if 99 out of 100 calls gave satisfactory answers, the printed
the minimum p-value (<0.05) and implied that the function was incorrect.

This has now been amended- we record how many of the p-values are <
0.05, and if this falls below 50%, then we print an error message
along with the minimum p-value. If however more than 50% pass, then we
print a success message, and the maximum p-value.