Allow optional non-printing of exp/actual values
`variable_error_messages._value_error` now has the `exp` and `res`
variables as optional parameters. All current calls to this internally
provide `exp` and `res`, so the behaviour won't change.

If `exp` and `res` are not provided, then the error message simply says
that the value is incorrect. This is in anticipation of a new top level
function which will check student-provided answers to symbolic questions
(i.e. which they calculate by hand). In that instance, printing the
expected answer is not desirable.