Allow permutations of answers
For some question types, it might be necessary for students to provide
answers that aren't strictly ordered. E.G reporting the eigenvalues of a
matrix. For these question types, I have now added the optional argument
`perm` to the check_answer function. Setting `perm=True` checks to make
sure that the student's answer matches with a permuatation of the
expected answer. Thus if the student sets

   x = [3, 2, 1]

Then
   check_vars('x', [1, 2, 3], perm=True)

will return True

In the call to `check_answer` this looks like

   check_answer('question_name', perm=True)